requirements.txt
openai>=1.0.0
deepseek>=0.0.1
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.13.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
pytest>=7.3.1
tqdm>=4.65.0
jsonschema>=4.17.0
scikit-learn>=1.2.0
transformers>=4.30.0
torch>=2.0.0
langchain>=0.0.300
python-dateutil>=2.8.2
rich>=13.0.0
